refactor(request): create main class by query string parameters
